requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
DateTime==5.2
Django==4.2.4
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
zope.interface==6.0
